ai/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
scikit-learn==1.4.2
pandas==2.2.2
numpy==1.26.4
joblib==1.4.0
pydantic==2.7.1
pymongo==4.7.0
python-dotenv==1.0.1
